fix: 방 페이지 path 변경 및 최소 너비 설정
