+++ client/src/app/room/[roomId]/style.module.scss
@@ -4,6 +4,9 @@
     display: flex;
     flex-flow: column nowrap;
     row-gap: 23px;
+    min-width: 1000px;
+    overflow-x: auto;
+    overflow-y: hidden;
 }
 
 .container {
