+++ client/src/styles/global.scss
@@ -1,3 +1,8 @@
+body {
+    height: 100%;
+    overflow-y: hidden;
+}
+
 a,
 button {
     cursor: pointer;
